don bang

+++ database/migrations/2024_02_12_095304_create_mahasiswas_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('mahasiswa', function (Blueprint $table) {
-            $table->integer('nim');
+            $table->bigInteger('nim');
             $table->unique('nim');
             $table->string('nama');
             $table->string('jurusan');
